moved nvcr_utility
